Refactor(jira): Move the AlertJiraIssueProperties to IssueProperties
